requirements.txt
langchain
mysql-connector-python
openai
uvicorn
fastapi
mysql.connector
pydantic
starlette